==> documents/modelling/02-Individuals.Rmd <==
```{r setupinds, include=FALSE}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(kableExtra)
if(knitr:::is_html_output()) options(knitr.table.format = "html") 
if(knitr:::is_latex_output()) options(knitr.table.format = "latex") 
library(tidyverse)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F)
library(ecoevosimulator)
```

# Ecoevosimulator

This chapter introduce the simulator with the different steps. Additional following chapters develop analyses that supported specific development choices or parameters.

Simulations start with **environment**, **genotypes** and **phenotypes** initialisation, and follows with a cycle of four steps repeated over time:

* **Forest gap dynamics**
* **Background mortality**
* **Reproduction**
* **Recruitment**

## Parameters summary

* $grid$ *int.* square matrix size
* $Nt$ *int* number of time steps
* $E_{lim}$ *double.* gradient maximal absolute value
* $\mu_G$ *double.* genetic mean (could be set to null)
* $\sigma^2_G$ *double.* genetic variance
* $\mu_E$ *double.* environmental mean (could be set to null)
* $\sigma^2_E$ *double.* environmental variance
* $P_{fall}$ *double.* treefall probability
* $R_{gaps}$ *int.* treefall gaps radius
* $P_{death}$ *double.* background mortality probability
* $N_s$ *int.* number of seedlings
* $R_{dispersal}$ *int.* dispersal distance
* $determinist$ *bool.* determinist (true) or probabilist (false) recruitment

## Initialisation

### Environment

The environmental mosaic where individuals grow is defined by a square matrix of size $grid\times grid$. 
Environment is generated with: (1) a gradient product, (2) a topography generator, or (3) a forest gap dynamics generator. Currently only the first method, a gradient product, is implemented. 
The environmental matrix varies in a grid defined by the product of the vectors $(−E_{lim},E_{lim})$ (Fig. \@ref(fig:env)).

```{r env, fig.cap="Example of environmental matrix obtained with gradient product."}
plotEnv(simulator(50,1))
```


### Trees

A tree grow in each cell of the environmental matrix. A tree $i$ is defined through its phenotype composed of a quantitative trait $z_i$ (two quantitative traits in next steps ?).
The quantitative trait $z_i$ is the sum of its genotypic breeding value $a_i$ and an environmental effect $e_i$, 
both drawn from a genetic and an environmental normal distribution of parameters $\mu_G$ and $\mu_E$ and $\sigma^2_G$ an $\sigma^2_E$,
the mean and the variance of genetic and environmental distributions (Fig. \@ref(fig:trees)):

\begin{equation} 
  z_i = a_i + e_i \\
  a_i \sim N(\mu_G,\sigma^2_G) \\
  e_i \sim N(\mu_e,\sigma^2_E)
  (\#eq:phenotype)
\end{equation} 

```{r trees, fig.cap="Example of genotype and phenotype matrices intialisation."}
plotMaps(simulator(50,1))
```

## Forest gap dynamics

**This one of the part to be improved, below is a quick implementation.**

The simulation occurs over $N_t$ time steps.
The time step is defined as the time between treefall events. 
We assume each cell to be filled by one mature canopy tree between two events.
We thus avoid the simulation of individual growth.
**But forest gap dynamics temporal and spatial structure should be studied with Paracou data.**

At each time step, every tree have a chance to fall with a probability $P_{fall}$.
Each fallen tree will create a gap of radius $R_{gaps}$.
All fallen trees, primary and secondary treefalls, are declared as dead and replaced by recruitment after reproduction.
Other trees undergo background mortality. 

## Background mortality

At each time step, every tree have a chance to die with a probability $P_{death}$.
This represent standing tree death.
Trees that survived forest gap dynamics and background mortality keep their genotypes and phenotypes for the next time step.

## Reproduction

$N_s$ seeds are dispersed in each empty cells due to treefall or background mortality.
Seedlings inherit their genotype from a mother at distance $R_{dispersal}$ from the empty cell and from a father at a distance $d$ from the mother (could be separated in $d_m$ and $d_f$).
The euclidean distance is explecitly computed, thus the dispersal kernel is a circle.
Parents are dranw from a uniform distribution in the dispersal kernel.
Seedlings genotype $a_s$ is drawn from a normal centered on parental mean and of half the variance of genetic variation $\sigma^2_G$, 
and is used to compute seedling trait value $z_s$ with environmental noise:

\begin{equation} 
  z_s = a_s + e_s \\
  a_s \sim N(mean(a_{P1},a_{P2}),\frac{\sigma^2_G}{2}) \\
  e_s \sim N(\mu_e,\sigma^2_E)
  (\#eq:as)
\end{equation} 

## Recruitment

One seedlings outcompete other $N_s-1$ seedlings with determinist or probabilist recruitment ($determinist$), 
and is recruited as a mature canopy tree for the next time step.
Viability is defined as the inverse of the euclidean distance between seedlings phenotype and the environment of the cell:

\begin{equation} 
  viability = \frac{1}{\sqrt{(z_s - e_i)^2}}
  (\#eq:viability)
\end{equation} 

Determinist recruitment select for the most viable (highest value) seedlings.
Probabilist recruitment use a random draw with viability as the recruitment probability.
