==> documents/modelling/03-ForestGapDynamics.Rmd <==
```{r setupfgd, include=FALSE}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(kableExtra)
if(knitr:::is_html_output()) options(knitr.table.format = "html") 
if(knitr:::is_latex_output()) options(knitr.table.format = "latex") 
library(tidyverse)
library(gganimate)
library(rstan)
library(bayesplot)
options(mc.cores = parallel::detectCores())
rstan_options(auto_write = T)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F)
library(ecoevosimulator)
```

# Forest gap dynamics

This chapter introduce the forest gap dynamics generator, mortality, and time-steps used in the simulator.
Generators and equation have been inferred from tree censuses in Paracou plots.
First, we simulated neighbourhood crowding index (NCI) dynamics in space and time.
Then, we inferred mortality and time-steps.
The forest gap dynamics is defined as a matrix spatially explicit in 2 dimensions and temporally explicit ($grid \times grid \times Nt~(time-steps)$).

## Neighbourhood crowding dynamics

We modeled neighbourhood crowding dynamics with a Markov process starting from an initial $NCI$ distribution,
and evolving at each iteration with a $\delta NCI$ variation.

### $NCI$

We computed the neighbourhood crowding index (NCI) for all trees in undisturbed plots (1, 6, 11 and 13 to 15) for all census years as follow:

\begin{equation} 
  NCI_{i,t} = \sum _{j~|~\delta_{i,j}<20m} ^{J_{i,t}} DBH_{j,t}^2 e^{-\frac14.\delta_{i,j}}
\end{equation} 

with $DBH_{j,t}$ the diameter from neighboring tree $j$ in year $t$ and $\delta_{i,j}$ its distance to individual tree $i$.
$NCI_{i,t}$ is computed for all neighbors at a distance $\delta_{i,j}$ inferior to maximum neighboring distance of 20 meters. 
The power of neighbors $DBH$ effect was set to 2 to consider neighbors surface.
The decrease of neighbors $DBH$ effect with distance was set to 0.25 here
to represent trees at 20 meters of the focal trees having 1% of the effect of the same tree at 0  meters.

The variation of plot 1 over time show only few decrease in NCI, that are related to falling trees (Fig. \@ref(fig:plotncigif)), 
as revealed by the NCI trajectories of the trees in plot 1 (Fig. \@ref(fig:ncip1)), 
showcasing regular increase of NCI with few stochastic decrease and a global stable mean.
$NCI$ followed a normal distribution centered on $\mu_{NCI} = 123$ of variance $\sigma_{NCI} = 26$ 
in all plot and censuses (Fig. \@ref(fig:p1spNCIdist)),
with spatially aggregated clusters on 10 meters or $d_{NCI} = 3$ cells (Fig. \@ref(fig:spcornci)).
We thus defined NCI distribution for initialization as:

\begin{equation} 
 NCI \sim \left\{
    \begin{array}{ll}
        \mathcal N(\mu_{NCI} = 123, \sigma_{NCI} = 26) \\
        d_{NCI}=3 & \mbox{Spatial aggregation}
    \end{array}
\right.
\end{equation} 

```{r, eval=F}
src_sqlite(file.path("/home/sylvain//Documents/BIOGECO/PhD/data/Paracou/",
                                "trees", "Paracou.sqlite")) %>% 
    tbl("Paracou") %>% 
    filter(CodeAlive == 1) %>% 
    filter(Plot %in% 1:15) %>%
    filter(CensusYear == 2017) %>% 
  group_by(Plot, SubPlot) %>% 
  select(Genus, Species) %>% 
  collect() %>% 
  unique() %>% 
  summarise(Nspecies = length(unique(Species)),
            Ngenus = length(unique(Genus))) %>% 
  mutate(Rha = Nspecies/Ngenus) %>% 
  ungroup() %>% 
  summarise_all(mean)
```



```{r ncidata, eval=F}
computeNCI <- function(CensusYear, Plot, d = 20){
  if(Plot == 16) 
    stop("Xfield and Yfield doesn't work with P16.")
  trees <- src_sqlite(file.path("/home/sylvain//Documents/BIOGECO/PhD/data/Paracou/",
                                "trees", "Paracou.sqlite")) %>% 
    tbl("Paracou") %>% 
    filter(CodeAlive == 1) %>% 
    filter(Plot == local(Plot)) %>% # test for P1
    filter(CensusYear == local(CensusYear)) %>% 
    mutate(DBH = CircCorr/pi) %>%
    collect()
  D <- dist(trees[c("Xfield", "Yfield")]) %>% 
    as.matrix()
  D[D>20] <- NA # individuals above 20 m
  D[D == 0] <- NA # theimselves
  D <- exp(-0.25*D)
  D[is.na(D)] <- 0
  trees$NCI <- as.vector(D %*% as.vector(trees$DBH))
  return(trees %>% 
           filter(Xfield > local(d), Xfield < 250-local(d), Yfield > local(d), Yfield < 250-local(d)))
}
nci <- lapply(unique(bkgmortality$Plot), function(p){
  lapply(unique(bkgmortality$CensusYear), 
              function(y) computeNCI(y, p)) %>% 
  bind_rows()
}) %>% 
  bind_rows()
write_tsv(nci, path = "modelling_save/nci.tsv")
```

```{r, eval=F}
library(raster)
nci <- read_tsv("modelling_save/nci.tsv") 
coordinates(nci) <- ~Xutm + Yutm
proj4string(nci) <- '+proj=utm +zone=22 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0'
wetness <- raster(file.path("/home/sylvain//Documents/BIOGECO/PhD/data/Paracou/",
                            "topography", "TWI_1m.tif"))
dem <- raster(file.path("/home/sylvain//Documents/BIOGECO/PhD/data/Paracou/",
                        "topography", "DEM_1m_2015.tif")) # for CRS
projection(wetness) <- projection(dem)
wetness <- projectRaster(wetness, crs = '+proj=utm +zone=22 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0')
nci$twi <- extract(wetness, nci)
as.data.frame(nci) %>% 
  ggplot(aes(twi, NCI)) +
  geom_point(alpha = 0.2) +
  geom_smooth(method = "lm") +
  ggpubr::stat_cor() +
  xlab("Topographic Wetness Index (TWI)") +
  ylab("Neighbour Crowding Index (NCI)")
```


```{r plotnci, eval=F}
g <- read_tsv("modelling_save/nci.tsv") %>% 
  mutate( CensusYear = as.integer(CensusYear)) %>% 
  filter(Plot %in% c(1)) %>% 
  ggplot(aes(Xfield, Yfield, col = log(NCI), group = idTree)) +
  geom_point() +
  facet_wrap(~ Plot, labeller = "label_both") +
  viridis::scale_color_viridis(direction = -1) +
  theme(axis.text = element_blank(), axis.title = element_blank(),
        axis.line = element_blank(), axis.ticks = element_blank()) +
   transition_time(CensusYear) +
  labs(title = 'Year: {frame_time}')
anim_save("modelling_save/nci.gif", g)
```

```{r plotncigif, fig.cap="NCI in the Plot 1 of Paracou."}
include_graphics("modelling_save/nci.gif")
```

```{r ncip1, fig.cap="NCI over census years in Plot 1."}
read_tsv("modelling_save/nci.tsv") %>% 
  filter(Plot == 6) %>%
  group_by(CensusYear) %>% 
  mutate(NCImean = mean(NCI)) %>% 
  ungroup() %>% 
  ggplot(aes(x = CensusYear)) +
  geom_line(aes(y = NCI, group = idTree), alpha = 0.1) +
  geom_line(aes(y = NCImean), col = "red", size = 2) +
  xlab("Census year") +
  ylab("neighbourhood crowding index (NCI) in Plot 1") 
```

```{r p1spNCIdist, fig.cap="Plot1 NCI distribution in space for all plots and censuses."}
t <-  read_tsv("modelling_save/nci.tsv") %>% 
  filter(Plot == 1, CensusYear == 2014) %>% 
  rename(X = Xfield, Y = Yfield)
mu <- MASS::fitdistr(t$NCI, "normal")$estimate[1]
sigma <- MASS::fitdistr(t$NCI, "normal")$estimate[2]
g1 <- ggplot(t, aes(x = NCI)) +
  geom_histogram(aes(y = ..density..), fill = "lightgrey", col = "grey", bins = 100) +
  stat_function(fun = dnorm , args = list(mean = mu, sd = sigma)) +
  geom_vline(xintercept = mu, linetype = "dashed") +
  geom_text(aes(y = y), label = "Normal", data = data.frame(NCI = 180, y = 0.015)) +
  geom_text(aes(y = y), label = expression(mu), data = data.frame(NCI = 180, y = 0.013)) +
  geom_text(aes(y = y), label = round(mu,3), data = data.frame(NCI = 180, y = 0.010)) +
  geom_text(aes(y = y), label = expression(sigma), data = data.frame(NCI = 180, y = 0.008)) +
  geom_text(aes(y = y), label = round(sigma,3), data = data.frame(NCI = 180, y = 0.005)) +
  xlab("NCI")
g2 <- ggplot(t, aes(X, Y, col = NCI)) +
  geom_point() +
  viridis::scale_color_viridis(direction = -1) +
  theme(axis.text = element_blank(), axis.title = element_blank(),
        axis.line = element_blank(), axis.ticks = element_blank())
cowplot::plot_grid(g1, g2)
```

```{r spcornci, fig.cap="Spatial autocorrelation of NCI for all plots and censuses."}
nci.cor <- pgirmess::correlog(t[c("X", "Y")], t$NCI, method = "Moran", nbclass = 100)
t <- as.data.frame(nci.cor) %>% 
  filter(dist.class < 20) %>% 
  mutate(coef = round(1000*coef)) %>% 
  group_by(dist.class) %>% 
  expand(coef = 1:coef)
lambdacor <- MASS::fitdistr(t$dist.class, "exponential")$estimate
nci.cor %>% 
  as.data.frame() %>% 
  filter(dist.class < 200) %>% 
  ggplot(aes(x = dist.class, y = coef)) + 
  geom_point(aes(alpha = p.value < 0.01)) + geom_line() +
  scale_alpha_manual("p < 0.01", values = c(0.3, 1)) +
  geom_vline(xintercept = 9, col = "black", linetype = "dashed") +
  xlab("Distance (m)") + ylab("Moran\'s I") +
  scale_color_discrete(expression(delta[NCI])) +
  stat_function(fun = function(x) exp(-lambdacor*x), col = "black") +
  geom_text(label = "Exponential", col = "black",
            data = data.frame(dist.class = 50, coef = 0.5, type = "positive")) +
  geom_text(label = expression(lambda),  col = "black",
            data = data.frame(dist.class = 50, coef = 0.4, type = "positive")) +
  geom_text(label = round(lambdacor,3),  col = "black",
            data = data.frame(dist.class = 50, coef = 0.3, type = "positive"))
```

### $\delta NCI$

We computed the NCI annual variation for all trees in undisturbed plots (1, 6, 11 and 13 to 15) for all censuses as follow:

\begin{equation} 
  \delta NCI_{i,census} = \frac{NCI_{i,census} - NCI_{i,census-1}}{t_{census} - t_{census-1}}
\end{equation} 

with $NCI_{i,census}$ and $NCI_{i,census-1}$ the neighbourhood crowding index in censuses at $t_{census}$ and $t_{census-1}$.
**Beware**, $t_{census} - t_{census-1}$ can be separated by more than one year, but the normalization gave us an annual $\delta NCI$.

$\delta NCI$ as a whole had a normal distribution inflated in zero (Fig. \@ref(fig:deltaNCIdistribution)).
We split the distribution of $\delta NCI$ between positive and negative values (Fig. \@ref(fig:deltaNCIdistribution)):

1. $\delta NCI>0$ represents trees in patches of growing trees following a lognormal distribution $\mu$ and $\sigma$.
1.  $\delta NCI \leq 0$ represents trees near treefalls following an exponential distribution of parameter $\lambda$ that deal with rare events of strong loss of NCI.

We used a Bernoulli model depending on previous $NCI$ to define the probability for $\delta NCI$ to be negative:

\begin{equation} 
  P(\delta NCI \leq 0) \sim \mathcal B(logit(\alpha + \beta.NCI_{census-1}))
\end{equation} 

We found significant non null $\alpha = -1.32$ and $\beta = 0.003$ (Fig. \@ref(fig:fideltaNCI)).

$\delta NCI>0$ and $\delta NCI \leq 0$ were both spatially aggregated in clusters of 10 meters or $d_{\delta NCI} = 3$ cells (Fig. \@ref(fig:spcor)).
We thus defined the distribution of NCI variation for the Markov process as:

\begin{equation} 
\delta NCI \sim \left\{
    \begin{array}{ll}
        P(\delta NCI \leq 0|\begin{array}{ll} \alpha=-1.32 \\ \beta=0.003 \end{array}) \sim \mathcal B(logit^{-1}(\alpha + \beta.NCI)) \\
        \mathcal {logN} (\mu_{\delta NCI > 0} = 0.749, \sigma_{\delta NCI > 0} = 2.651) & \mbox{if } \delta NCI > 0 \\
        \mathcal - {Exp} (\lambda_{\delta NCI \leq 0} = 0.31) & \mbox{ifelse} \\
        d_{\delta NCI}=3 & \mbox{Spatial aggregation}
    \end{array}
\right.
\end{equation} 

```{r deltaNCIdistribution, fig.cap="Delta NCI distribution per year in plots 1, 6, 11 and 13 to 15."}
nci <- read_tsv("modelling_save/nci.tsv") %>% 
  group_by(idTree) %>% 
  arrange(CensusYear) %>% 
  mutate(deltaNCI = (NCI - lag(NCI))/(CensusYear - lag(CensusYear)),
         NCIlag = lag(NCI)) %>% 
  ungroup() %>% 
  na.omit()
nci.pos <- filter(nci, deltaNCI > 0)
nci.neg <- filter(nci, deltaNCI <= 0)
p <- nrow(nci.neg)/nrow(nci)
mu <- MASS::fitdistr(nci$deltaNCI, "normal")$estimate[1]
sigma <- MASS::fitdistr(nci$deltaNCI, "normal")$estimate[2]
mulog <- MASS::fitdistr(nci.pos$deltaNCI, "lognormal")$estimate[1]
sigmalog <- MASS::fitdistr(nci.pos$deltaNCI, "lognormal")$estimate[2]
lambda <- MASS::fitdistr(-filter(nci, deltaNCI <= 0)$deltaNCI, "exponential")$estimate
g.deltaNCI <- ggplot(nci, aes(x = deltaNCI)) +
  geom_histogram(aes(y = ..density..), fill = "lightgrey", col = "grey", bins = 100)  +
  stat_function(fun = dnorm , args = list(mean = mu, sd = sigma)) +
  geom_vline(xintercept = mu, linetype = "dashed") +
  geom_text(aes(y = y), label = "Normal", data = data.frame(deltaNCI = 30, y = 0.25)) +
  geom_text(aes(y = y), label = expression(mu), data = data.frame(deltaNCI = 30, y = 0.2)) +
  geom_text(aes(y = y), label = round(mu,3), data = data.frame(deltaNCI = 30, y = 0.15)) +
    geom_text(aes(y = y), label = expression(sigma), data = data.frame(deltaNCI = 30, y = 0.10)) +
  geom_text(aes(y = y), label = round(sigma,3), data = data.frame(deltaNCI = 30, y = 0.05)) +
  xlab(expression(delta[NCI]))
g.deltaNCI.positive <- ggplot(nci.pos, aes(x = deltaNCI)) +
  geom_histogram(aes(y = ..density..), fill = "lightgrey", col = "grey", bins = 100)  +
  stat_function(fun = dlnorm , args = list(meanlog = mulog, sdlog = sigmalog)) +
  geom_vline(xintercept = exp(mu), linetype = "dashed") +
  geom_text(aes(y = y), label = "Logormal", data = data.frame(deltaNCI = 20, y = 0.5)) +
  geom_text(aes(y = y), label = expression(mu), data = data.frame(deltaNCI = 20, y = 0.4)) +
  geom_text(aes(y = y), label = round(exp(mulog),3), data = data.frame(deltaNCI = 20, y = 0.3)) +
    geom_text(aes(y = y), label = expression(sigma), data = data.frame(deltaNCI = 20, y = 0.2)) +
  geom_text(aes(y = y), label = round(exp(sigmalog),3), data = data.frame(deltaNCI = 20, y = 0.1)) +
  xlab(expression(delta[NCI]>0))
g.deltaNCI.negative <- ggplot(nci.neg, aes(x = -deltaNCI)) +
  geom_histogram(aes(y = ..density..), fill = "lightgrey", col = "grey", bins = 100)  +
  stat_function(fun = dexp , args = list(rate = lambda)) +
  geom_vline(xintercept = lambda, linetype = "dashed") +
  geom_text(aes(y = y), label = "Exponential", data = data.frame(deltaNCI = -50, y = 0.25)) +
  geom_text(aes(y = y), label = expression(lambda), data = data.frame(deltaNCI = -50, y = 0.2)) +
  geom_text(aes(y = y), label = round(lambda,3), data = data.frame(deltaNCI = -50, y = 0.15)) +
  xlab(expression(-delta[NCI] <= 0))
cowplot::plot_grid(g.deltaNCI, g.deltaNCI.negative, g.deltaNCI.positive, nrow = 3)
```

```{r fideltaNCI, fig.cap="$\\alpha$ and $\\beta$ posteriors for the risk of $\\delta NCI$$ to be negative with uncentered data and 10^4 observations.", fig.height=4}
# mdata <- nci %>% 
#   mutate(NCIdecrease = as.integer(deltaNCI <= 0)) %>% 
#   sample_n(10^4)
# model <- stan_model(file = "modelling_stan/deltaNCI.stan")
# fit <- sampling(model, data = list(N = nrow(mdata),
#                                    NCIdecrease = mdata$NCIdecrease,
#                                    NCI = mdata$NCIlag))
# mcmc_trace(fit, np = nuts_params(fit))
# mcmc_pairs(fit, np = nuts_params(fit))
# save(mdata, fit, file = "modelling_save/fitdeltaNCI.stan")
load("modelling_save/fitdeltaNCI.stan")
mcmc_dens(fit, pars = c("alpha", "beta"), facet_args = list())
```

```{r spcor, fig.cap="Spatial autocorrelation of deltaNCI."}
# nci.pos.2015 <- filter(nci.pos, CensusYear == 2015) %>% 
#   filter(Plot == 1)
# nci.pos.2015.cor <- pgirmess::correlog(nci.pos.2015[c("Xfield", "Yfield")], nci.pos.2015$deltaNCI, method = "Moran", nbclass = 100)
# nci.neg.2015 <- filter(nci.neg, CensusYear == 2015) %>% 
#   filter(Plot == 1)
# nci.neg.2015.cor <- pgirmess::correlog(nci.neg.2015[c("Xfield", "Yfield")], nci.neg.2015$deltaNCI, method = "Moran", nbclass = 100)
# nci.cor <- bind_rows(mutate(as.data.frame(nci.pos.2015.cor), type = "positive"),
#                      mutate(as.data.frame(nci.neg.2015.cor), type = "negative"))
# write_tsv(nci.cor, "modelling_save/spcor.tsv")
t <- read_tsv("modelling_save/spcor.tsv") %>% 
  filter(dist.class < 20) %>% 
  mutate(coef = round(1000*coef)) %>% 
  group_by(dist.class) %>% 
  expand(coef = 1:coef)
lambdacor <- MASS::fitdistr(t$dist.class, "exponential")$estimate
read_tsv("modelling_save/spcor.tsv") %>% 
  as.data.frame() %>% 
  filter(dist.class < 200) %>% 
  ggplot(aes(x = dist.class, y = coef, col = type)) + 
  geom_point(aes(alpha = p.value < 0.01)) + geom_line() +
  scale_alpha_manual("p < 0.01", values = c(0.3, 1)) +
  geom_vline(xintercept = 9, col = "black", linetype = "dashed") +
  xlab("Distance (m)") + ylab("Moran\'s I") +
  scale_color_discrete(expression(delta[NCI])) +
  stat_function(fun = function(x) exp(-lambdacor*x), col = "black") +
  geom_text(label = "Exponential", col = "black",
            data = data.frame(dist.class = 50, coef = 0.5, type = "positive")) +
  geom_text(label = expression(lambda),  col = "black",
            data = data.frame(dist.class = 50, coef = 0.4, type = "positive")) +
  geom_text(label = round(lambdacor,3),  col = "black",
            data = data.frame(dist.class = 50, coef = 0.3, type = "positive"))
```

### Generator

Using the equations defining NCI for initialization and NCI variation for the Markov process, 
we obtained the following definition of NCI dynamics:

\begin{equation} 
 NCI \sim \left\{
    \begin{array}{ll}
        \mathcal N(\mu_{NCI} = 123, \sigma_{NCI} = 26) \\
        d_{NCI}=3 & \mbox{Spatial aggregation}
    \end{array}
\right.
\end{equation} 

\begin{equation} 
\delta NCI \sim \left\{
    \begin{array}{ll}
        P(\delta NCI \leq 0|\begin{array}{ll} \alpha=-1.32 \\ \beta=0.003 \end{array}) \sim \mathcal B(logit^{-1}(\alpha + \beta.NCI)) \\
        \mathcal {logN} (\mu_{\delta NCI > 0} = 0.749, \sigma_{\delta NCI > 0} = 2.651) & \mbox{if } \delta NCI > 0 \\
        \mathcal {Exp} (\lambda_{\delta NCI \leq 0} = 123) & \mbox{ifelse} \\
        d_{\delta NCI}=3 & \mbox{Spatial aggregation}
    \end{array}
\right.
\end{equation} 

We simulated back $NCI$ and $\delta NCI$ dynamics,
using random generators based on probabilistic laws,
and disaggregating a grid to the desired size of spatial aggregation ($d=3$). 
Using inferred parameters we successfully simulated back $NCI$ 
and $\delta NCI$ spatial distribution (Fig. \@ref(fig:simspdNCIdist) and Fig. \@ref(fig:simspNCIdist)).
We then initialized $NCI$ with $NCI$ generator and simulated its dynamics variation over 33 years using $\delta NCI$ generator.
We obtain satisfying variation over time and space (Fig. \@ref(fig:simulatedNCI) and Fig. \@ref(fig:simncigif)).
We are thus able to simulated $NCI$ spatio-temporal dynamics based on **8** parameters,
using only one $d$ parameter for both $NCI$ and $\delta NCI$ spatial structure.

\begin{equation} 
NCI_{(x,y),t} = f 
  \begin{pmatrix}
    \mu_{NCI} \\ 
    \sigma_{NCI} \\ 
    \alpha \\ 
    \beta \\ 
    \mu_{\delta NCI > 0} \\ 
    \sigma_{\delta NCI > 0} \\ 
    \lambda_{\delta NCI \leq 0} \\ 
    d
  \end{pmatrix}
\end{equation} 

```{r rdeltanci}
rdeltanci <- function(
  NCI = rnorm(1, mean = 122.73, sd = 25.806),
  alpha = -1.32,
  beta = 0.003,
  mu = 0.749, # mu of lognormal distributon for positive deltanci
  sigma = 2.651, # sigma of lognormal distributon for positive deltanci
  lambda = 0.31 # lamnda of exponential distributon for negative deltanci
){
  samples <- rep(0, length(NCI))
  P <- sapply(arm::invlogit(alpha + beta*NCI), function(p) rbinom(1,1,p)) == 1
  samples[!P] <- rlnorm(sum(!P), meanlog = log(mu), sdlog = log(sigma))
  samples[P] <- -rexp(sum(P), rate = lambda)
  return(samples)
}
```

```{r rspdeltanci}
rspcor <- function(
  M =  matrix(rdeltanci(NCI = rnorm(50*50, mean = 122.73, sd = 25.806)), 
              ncol = 50, nrow = 50),
  dcor = 3 # spatial autocorrelation size (3*3m)
){
  M <- raster::raster(M)
  M <- raster::aggregate(M, dcor)
  M <- raster::disaggregate(M, dcor, method = "bilinear")
  raster::as.matrix(M)[1:50,1:50] %>% 
    reshape2::melt() %>% 
    rename(X = Var1, Y = Var2)
} 
```

```{r simspdNCIdist, fig.cap="Simulated deltaNCI distribution in space for a random time."}
simnci <- data.frame(NCI = rnorm(nrow(nci)*10, mean = 122.73, sd = 25.806)) %>% 
  mutate(deltaNCI = rdeltanci(NCI))
g1 <- ggplot(simnci, aes(x = deltaNCI)) +
  geom_histogram(aes(y = ..density..), fill = "lightgrey", col = "grey", bins = 100)  +
  stat_function(fun = dnorm , args = list(mean = mean(simnci$deltaNCI), 
                                          sd = sd(simnci$deltaNCI))) +
  geom_vline(xintercept = mean(simnci$deltaNCI), linetype = "dashed") +
  geom_text(aes(y = y), label = "Normal", data = data.frame(deltaNCI = 30, y = 0.25)) +
  geom_text(aes(y = y), label = expression(mu), data = data.frame(deltaNCI = 30, y = 0.2)) +
  geom_text(aes(y = y), label = round(mean(simnci$deltaNCI),3), data = data.frame(deltaNCI = 30, y = 0.15)) +
    geom_text(aes(y = y), label = expression(sigma), data = data.frame(deltaNCI = 30, y = 0.10)) +
  geom_text(aes(y = y), label = round(sd(simnci$deltaNCI),3), data = data.frame(deltaNCI = 30, y = 0.05)) +
  xlab(expression(delta[NCI])) 
g2 <- ggplot(rspcor(matrix(simnci$deltaNCI[1:(50*50)], ncol = 50, nrow = 50)), 
             aes(X, Y, fill = value)) +
  geom_tile() +
  viridis::scale_fill_viridis(expression(delta[NCI]), direction = -1) +
  theme(axis.text = element_blank(), axis.title = element_blank(),
        axis.line = element_blank(), axis.ticks = element_blank())
cowplot::plot_grid(g1, g2)
```

```{r simspNCIdist, fig.cap="Simulated NCI distribution in space for a random time."}
t <- rspcor(matrix(rnorm(50*50, mean = 122.73, sd = 25.806), nrow = 50, ncol = 50), dcor = 4)
g1 <- ggplot(t, aes(x = value)) +
  geom_histogram(aes(y = ..density..), fill = "lightgrey", col = "grey", bins = 100) +
  xlab(expression(NCI))
g2 <- ggplot(t, aes(X, Y, fill = value)) +
  geom_tile() +
  viridis::scale_fill_viridis(expression(NCI), direction = -1) +
  theme(axis.text = element_blank(), axis.title = element_blank(),
        axis.line = element_blank(), axis.ticks = element_blank())
cowplot::plot_grid(g1, g2)  
```

```{r simncitsv, eval=F}
t <- rspcor(matrix(rnorm(50*50, mean = 122.73, sd = 25.806), nrow = 50, ncol = 50), dcor = 4) %>% 
  mutate(CensusYear = 1984) %>% 
  rename(NCI = value)
for(y in 1985:2017)
  t <- bind_rows(t, filter(t, CensusYear == y - 1) %>% 
                   mutate(NCI = NCI + rspcor(matrix(rdeltanci(NCI), 
                                                    nrow = 50, ncol = 50))$value,
                          CensusYear = CensusYear + 1)) 
write_tsv(t, path = "modelling_save/simnci.tsv")
```

```{r simulatedNCI, fig.cap="Simulated NCI over census years."}
read_tsv("modelling_save/simnci.tsv", col_types = cols(deltaNCI = col_double())) %>% 
  mutate(idTree = paste0(X,"-",Y)) %>% 
  group_by(CensusYear) %>% 
  mutate(NCImean = mean(NCI)) %>% 
  ungroup() %>% 
  ggplot(aes(x = CensusYear)) +
  geom_line(aes(y = NCI, group = idTree), alpha = 0.1) +
  geom_line(aes(y = NCImean), col = "red", size = 2) +
  xlab("Census year") +
  ylab("Simulated neighbourhood crowding index in time.")  
```

```{r simnci, eval=F}
g <- read_tsv("modelling_save/simnci.tsv", col_types = cols(deltaNCI = col_double())) %>% 
  mutate( CensusYear = as.integer(CensusYear)) %>% 
  ggplot(aes(X, Y, fill = log(NCI))) +
  geom_tile() +
  viridis::scale_fill_viridis(direction = -1) +
  theme(axis.text = element_blank(), axis.title = element_blank(),
        axis.line = element_blank(), axis.ticks = element_blank()) +
  transition_time(CensusYear) +
  labs(title = 'Year: {frame_time}')
anim_save("modelling_save/simnci.gif", g)
```

```{r simncigif, fig.cap="Simulated NCI."}
include_graphics("modelling_save/simnci.gif") 
```

## Mortality

We used a Bernoulli model to define the probability of death per year $P(death)$ (Fig. \@ref(fig:plot1mortalitygif)),
based on a $year$ random effect (e.g. climatic effect as El Nino events) 
and tree $DBH$, $NCI$ and $\delta NCI$ (Fig. \@ref(fig:mortality)):

\begin{equation} 
 P(death_{year}) \sim \mathcal B(logit(\theta_{0,year}+\theta_{DBH}. DBH+\theta_{NCI}.NCI+\theta_{\delta NCI}.\delta  NCI)) \\
 \theta_{0,year} \sim \mathcal N(\theta_0, \sigma_{year})
\end{equation} 

We considered no fixed effects as significant (Fig. \@ref(fig:fitmortality)),
and thus obtained the following equation to simulate mortality:

\begin{equation} 
 P(death) \sim \mathcal B(p = 0.01325208)
\end{equation} 

```{r plot1mortality, eval=F}
g <- src_sqlite(file.path("/home/sylvain//Documents/BIOGECO/PhD/data/Paracou/",
                     "trees", "Paracou.sqlite")) %>%
  tbl("Paracou") %>%
  filter(Plot == 1) %>%
  collect() %>% 
  mutate_at(c("CodeAlive", "CodeMeas"), as.factor) %>% 
  mutate(gaps = as.factor(as.numeric(CodeMeas %in% 6:7 & CodeAlive == 0))) %>% 
  mutate(background = as.factor(as.numeric(CodeMeas == 0 & CodeAlive == 0))) %>% 
  mutate(death = "alive") %>% 
  mutate(death = ifelse(gaps == 1, "gaps", death)) %>% 
  mutate(death = ifelse(background == 1, "background", death)) %>% 
  ggplot(aes(Xfield, Yfield, col = death, alpha = death, group = idTree)) +
  geom_point() +
  scale_color_manual("Mortality", values = c("lightgrey", "red", "blue")) +
  scale_alpha_discrete("Mortality") +
  facet_wrap(~ Plot, labeller = "label_both") +
  theme(axis.text = element_blank(), axis.title = element_blank(),
        axis.line = element_blank(), axis.ticks = element_blank()) +
  transition_time(CensusYear) +
  labs(title = 'Year: {frame_time}')
anim_save("modelling_save/plot1treefalls.gif", g)
```
```{r plot1mortalitygif, fig.cap="Mortality in the Plot 1 of Paracou. In blue death on foot with code measure 0 and in red death by treefalls with code measure 6 and 7."}
include_graphics("modelling_save/plot1treefalls.gif")
```

```{r mortality, fig.cap="Mortality for all censuses in Paracou per explanatory variables."}
nci <- read_tsv("modelling_save/nci.tsv") %>% 
  left_join(src_sqlite(file.path("/home/sylvain//Documents/BIOGECO/PhD/data/Paracou/",
                     "trees", "Paracou.sqlite")) %>%
  tbl("Paracou") %>%
  dplyr::select(CensusYear) %>% 
  collect() %>% 
  unique() %>% 
  arrange(CensusYear) %>% 
  mutate(NextCensusYear = lead(CensusYear))) %>%  # current computed is the NCI before the death of individual in the enxt census
  dplyr::select(idTree, NextCensusYear, NCI) %>% 
  na.omit() %>% 
  rename(CensusYear = NextCensusYear) %>% 
  group_by(idTree) %>% 
  arrange(CensusYear) %>% 
  mutate(deltaNCI = (NCI - lag(NCI))/(CensusYear - lag(CensusYear)),
         NCIlag = lag(NCI)) %>% 
  ungroup() %>% 
  na.omit()
data <- src_sqlite(file.path("/home/sylvain//Documents/BIOGECO/PhD/data/Paracou/",
                     "trees", "Paracou.sqlite")) %>%
  tbl("Paracou") %>%
  filter(Plot %in% c(1, 6, 11, 13:15)) %>% 
  filter(CodeAlive == 1 | (CodeAlive == 0 & CodeMeas %in% c(0,6:7))) %>% 
  mutate(DBH = CircCorr/pi) %>% 
  collect() %>% 
  left_join(nci) %>% 
  na.omit() %>% 
  mutate(Code = paste0(CodeAlive, CodeMeas)) %>%
  mutate(Status = recode(Code,
                            "00" = "background",
                            "06" = "primary treefall",
                            "07" = "secondry treefall",
                            "10" = "alive",
                            "11" = "alive",
                            "12" = "alive",
                            "13" = "alive",
                            "14" = "alive",
                            "17" = "alive")) %>% 
  mutate(CensusYearInteger = as.integer(as.factor(CensusYear))) %>% 
  mutate(death = as.integer(Status != "alive"))
data %>% 
  dplyr::select(Status, CensusYear, DBH, NCI, deltaNCI) %>% 
  reshape2::melt(id.vars = "Status") %>% 
  group_by(variable) %>% 
  mutate(valueclass = cut(value, breaks = quantile(value, probs = seq(0, 1, 0.2)), 
                          include.lowest = T)) %>% 
  group_by(Status, variable, valueclass) %>% 
  summarise(N = n()) %>% 
  reshape2::dcast(variable + valueclass ~ Status) %>% 
  reshape2::melt(id.vars = c("variable", "valueclass", "alive"), 
                 variable.name = "type", value.name = "N") %>% 
  mutate(mortality = N/alive) %>% 
  ggplot(aes(valueclass, mortality, fill = type)) +
  geom_col(position = "dodge") +
  coord_flip() +
  facet_wrap(~ variable, scales = "free") +
  theme(legend.position = "bottom", axis.title = element_blank())
```

```{r fitmortality, fig.cap="Parameters posteriors for the tree mortality model with centered data and 10^4 observations."}
# mdata <- sample_n(data, 10^4)
# model <- stan_model(file = "modelling_stan/mortality.stan")
# fit <- sampling(model, data = list(N = nrow(mdata),
#                                    Y = max(mdata$CensusYearInteger),
#                                    death = mdata$death,
#                                    year = mdata$CensusYearInteger,
#                                    DBH = as.vector(scale(mdata$DBH)),
#                                    NCI = as.vector(scale(mdata$NCI)),
#                                    deltaNCI = as.vector(scale(mdata$deltaNCI))),
#                 include = F, pars = c("theta0y"))
# mcmc_trace(fit, np = nuts_params(fit))
# mcmc_pairs(fit, np = nuts_params(fit))
# save(mdata, fit, file = "modelling_save/fitmortality0.stan")
load("modelling_save/fitmortality0.stan")
mcmc_dens(fit, regex_pars = "theta")
```

## Time steps

To avoid simulating tree growth and DBH,
we simulated mature individuals (reproducing) in the canopy.
Following @Hardy2006, *Symphonia globulifera* reach maturity at $DBH_{maturity} = 17~cm$ (minimum adult DBH),
and based on annual growth rate parameters *Symphonia globulifera* has a liberation diameter of $Dopt=25~cm$ 
([symcapture analyses](https://phdpages.netlify.app/6b189d4142e8224bff99e28abd11cbfdd50c51b1/symcapture/growth-genomics.html)).
We thus used a maturity DBH of $DBH_{maturity} = 20~cm$.
Then we related growth to time with annual growth rate parameters
$Gmax=0.53~cm.yr^{-1}$, $Dopt=25~cm$, and $Ks=0.70$ 
([symcapture analyses](https://phdpages.netlify.app/6b189d4142e8224bff99e28abd11cbfdd50c51b1/symcapture/growth-genomics.html)),
despite evidences for strong differences among individuals.
We found *Symphonia* individuals to reach maturity between 28 and 41 years (Fig. \@ref(fig:dbhintime)).
Consequently, we decided to use a **time step of 30 years**.
This time-step corresponds to a mean mortality of 32% of individuals at each step (Fig. \@ref(fig:mortalityovertime)).

```{r dbhintime, fig.cap="Age at maturity since recruitment per species."}
agr <- function(dbh, gmax = 0.53, dopt = 25, Ks = 0.7)
  gmax*exp(-0.5*(log(dbh/dopt)/Ks)^2)
dbhintime <- function(ymax = 100, gmax = 0.53, dopt = 25, Ks = 0.7){
  censuses <- data.frame(year = 1, dbh = 10) %>% 
    mutate(agr = agr(dbh, gmax = gmax, dopt = dopt, Ks = Ks))
  for(t in 2:100){
    census <- filter(censuses, year == (t - 1)) %>% 
      mutate(dbh = dbh + agr) %>% 
      mutate(agr = agr(dbh, gmax = gmax, dopt = dopt, Ks = Ks)) %>% 
      mutate(year = t)
    censuses <- bind_rows(censuses, census)
  }  
  return(censuses)
}
lapply(list("S. sp.2" = 0.537, "S. globulifera" = 0.542, "S. sp.1" = 0.367), 
       function(g) dbhintime(gmax = g)) %>% 
  bind_rows(.id = "Species") %>% 
  # group_by(Species) %>% filter(dbh >= 20) %>%  filter(dbh == first(dbh))
  ggplot(aes(year, dbh, col = Species)) +
  geom_line() +
  geom_hline(yintercept = 20, linetype = "dashed") +
  geom_vline(aes(xintercept = year, col = Species), linetype = "dashed",
             data.frame(year = c(28, 28, 41), 
                        Species = c("S. sp.2", "S. globulifera", "S. sp.1"))) +
  xlab("Age since recruitment (year)") + ylab("DBH (cm)") +
  geom_text(aes(label = label), 
            data = data.frame(year = c(20, 50), dbh = 25, label = c(28, 41), 
                              Species = c("S. globulifera", "S. sp.1")))
```

```{r mortalityovertime, fig.cap="Death probability integrated over time."}
data.frame(Years = 1:500) %>% 
  mutate(Pdeath = 1 - dbinom(0, Years, 0.013)) %>% 
  ggplot(aes(Years, Pdeath)) + 
  geom_line() + 
  geom_vline(xintercept = 1:15*30, col = "lightgrey") + 
  geom_hline(yintercept = 0.32467227, col = "red") +
  scale_x_sqrt() +
  ggtitle("P(death over 30 years | Pdeath = 0.013) = B(30, Pdeath) = 0.32467227")
```
