==> documents/modelling/04-Ecoevosimulator.Rmd <==
```{r setupsim, include=FALSE}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(kableExtra)
if(knitr:::is_html_output()) options(knitr.table.format = "html") 
if(knitr:::is_latex_output()) options(knitr.table.format = "latex") 
library(tidyverse)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F)
library(ecoevosimulator)
```

# Ecoevosimulator

This chapter introduces the the core of the simulator.
It uses topography and NCI matrices generated in previous chapters.

Simulation follows three basic steps for each tree in each cell of the $grid \times grid$ matrix over $Nt$ time-steps (Fig. \@ref(fig:simsim)):

1. **Death**: The tree risks dying. Dead trees are replaced with reproduction and recruitment. Alive trees form the overlapping generations.
1. **Reproduction**: Seedlings are produced in each empty cell.
1. **Recruitment**: One seedling is recruited as a mature tree in canopy among seedlings in each empty cell.

```{r simsim, fig.cap="Scheme of the EcoEvoSimulator with death, reproduction and recruitment steps. See text above."}
knitr::include_graphics("images/EcoEvoSimulator.png")
```

We additionally precised the initialization of the model using random draws,
that might further be modified to represent specific starts.

## Death

At each time step and in each cell, 
the tree risks dying with a random draw in a binomial distribution:

\begin{equation} 
  Death \sim \mathcal B(timestep, Pdeath)
\end{equation} 

where $timestep$ (year) is the number of Bernoulli trials and $Pdeath$ is the annual death probability.
Dead trees are replaced with reproduction and recruitment steps. 
Alive trees form the overlapping generations.

## Reproduction

At each time step and in each empty cell,
$N_s$ seedlings are produced with trees from the previous time-step (including the dead tree),
using a mother at distance $R_{dispersion}$ from the empty cell and a father at distance $R_{pollination}$ from the mother.
Dispersal kernels are discrete circles with uniform probabilities and include the current cells (auto-fecundation).
Seedlings genotypes $g_s$ are drawn from a normal distribution centered on parental mean genotype with variance of half of genetic variance $\sigma^2_G$ [@Vincenzi2014].
Seedlings phenotypes $p_s$ are drawn from a normal distribution centered on seedlings genotypes $g_s$ with plastic variance $\sigma^2_P$:

\begin{equation} 
  p_s \sim \mathcal N(a_s,\sigma^2_P) \\
  g_s \sim \mathcal N(\frac{a_{P1}+a_{P2}}{2},\frac{\sigma^2_G}{2})
\end{equation} 

Note that two genotypes and phenotypes are defined for seedlings, both for topography and forest gap dynamics.

## Recruitment

At each time step and in each empty cell,
one seedling is recruited as a mature tree in canopy among the $N_s$ seedlings.
Seedlings are recruited with deterministic or probabilistic recruitment depending on their viability.
Deterministic recruitment selects the successful seedling with the highest viability.
Probabilistic recruitment draws the successful seedling with a probability proportional to seedlings viability.
Viability is defined as the inverse of the Euclidean distance between seedlings phenotype $p_s$ and the environment $e_s$ reduced by environment variance $\sigma^2_E$:

\begin{equation} 
  viability_s = \frac{1}{\sqrt{(\frac{p_s - e_s}{\sigma^2_E})^2}}
\end{equation} 

Note that the Euclidean distance used for viability is calculated in two dimensions composed by topography and forest gap dynamics.

## Initialisation

Currently, the initialization of the model uses random draws, 
but might further be modified to represent specific initialization.

In each cell at the fist step, 
a tree is drawn with a genotype $g_i$ drawn from a normal distribution centered on environmental mean $\mu_E$ with a genetic variance $\sigma^2_G$
and a phenotype $p_i$ drawn from a normal distribution centered $g_i$ with a plastic variance $\sigma^2_P$:

\begin{equation} 
  p_i \sim \mathcal N(g_i, \sigma^2_E) \\
  g_i \sim N(\mu_E, \sigma^2_G)
\end{equation} 

Note that two genotypes and phenotypes are defined for trees, both for topography and forest gap dynamics.
